# Faster dependency copying